+++ src/components/LodgingDetail/index.js
@@ -1,0 +1,3 @@
+export { default as LodgingDetailHeader } from './LodgingDetailHeader';
+export { default as LodgingDetailContent } from './LodgingDetailContent';
+export { default as LodgingDetailReview } from './LodgingDetailReview';
